# Conflicts:
#	public/build/js/app-vue-42ab0fac3e.js
#	public/build/js/app-vue-8f57eb52d7.js
#	public/build/js/app-vue-8f89e89623.js
#	public/build/js/app-vue.js.map
#	public/build/rev-manifest.json
#	public/js/app-vue.js
#	public/js/app-vue.js.map
#	resources/assets/js/vue-components.js
#	resources/views/control_presupuesto/cambio_presupuesto/create.blade.php

+++ resources/views/control_presupuesto/cambio_presupuesto/variacion_volumen.blade.php
@@ -75,7 +75,7 @@
                                     <td>@{{ concepto.sector }}</td>
                                     <td>@{{ concepto.cuadrante }}</td>
                                     <td>@{{ concepto.unidad }}</td>
-                                    <td class="text-right">@{{ parseInt(concepto.cantidad_presupuestada).formatMoney(2, ',','.') }}</td>
+                                    <td class="text-right">@{{ parseFloat(concepto.cantidad_presupuestada).formatMoney(2, ',','.') }}</td>
                                     <td>
                                         <div class="form-group" :class="{'has-error': validation_errors.has('form_save_solicitud.Variacion de volumen ' + (i+1))}">
                                             <input type="text" :name="'Variacion de volumen ' + (i+1)" v-validate="'required|decimal|regex:[^0]+'" class="form-control input-sm" v-model="concepto.variacion_volumen">
